refactor generate_unique_username to lib
